Separate functions into individual modules
* Remove babel-plugin-add-module-exports dependency
* Remove "internal" functions from the public API
* Fix tree-shaking support by exposing functions individually
