+++ .babelrc
@@ -14,9 +14,6 @@
       ]
     },
     "commonjs": {
-      "plugins": [
-        "add-module-exports"
-      ],
       "presets": [
         [
           "env",
