+++ package.json
@@ -55,7 +55,6 @@
     "babel-cli": "^6.24.1",
     "babel-core": "^6.24.1",
     "babel-loader": "^7.0.0",
-    "babel-plugin-add-module-exports": "^0.2.1",
     "babel-plugin-istanbul": "^4.1.4",
     "babel-preset-env": "^1.5.2",
     "babel-register": "^6.24.1",
