+++ src/internal/axiosInstance.js
@@ -1,0 +1,13 @@
+import Axios from 'axios';
+
+/**
+ * The Axios instance used for API queries. Not meant for general use.
+ *
+ * @private
+ */
+export default Axios.create({
+  baseURL: 'https://haveibeenpwned.com/api',
+  headers: {
+    Accept: 'application/vnd.haveibeenpwned.v2+json',
+  },
+});
